# docs/openapi/identitycore-public-api.yaml
openapi: 3.1.0
info:
  title: IdentityCore Public API
  version: 0.3.0
  description: Public REST API surface plus workspace-authenticated portal routes used by IdentityCore SDKs and sandbox testing.
servers:
  - url: https://api.identitycore.com/api/v1
    description: Production
  - url: http://localhost:8000/api/v1
    description: Local development
security:
  - apiClient: []
components:
  securitySchemes:
    apiClient:
      type: http
      scheme: bearer
      bearerFormat: API client secret
      description: Send the API client secret as a bearer token and the client ID in `X-Client-Id`.
    platformUserSession:
      type: apiKey
      in: cookie
      name: sessionid
      description: Portal session authentication used by authenticated workspace users.
    platformUserBearer:
      type: http
      scheme: bearer
      description: Portal JWT authentication used by authenticated workspace users.
  parameters:
    ClientIdHeader:
      name: X-Client-Id
      in: header
      required: true
      schema:
        type: string
      description: API client ID.
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 255
      description: A caller-generated key retained for 24 hours. Reusing it with the same payload replays the original result.
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [success, error, request_id]
      properties:
        success:
          const: false
        error:
          $ref: "#/components/schemas/Error"
        request_id:
          type: string
...
    VerificationCreateRequest:
      type: object
      required: [purpose, policy_id, verification_subject]
      properties:
        external_reference:
          type: string
        purpose:
          type: string
        policy_id:
          type: string
          description: Active verification policy/template ID. Required for API-client requests.
        project_id:
          type: string
          description: Project environment identifier. Defaults to the tenant's sandbox project for backward compatibility.
        verification_subject:
          $ref: "#/components/schemas/VerificationSubjectInput"
        redirect_url:
          type: string
          format: uri
        metadata:
          type: object
          additionalProperties: true
    VerificationCreateResponse:
      type: object
      required: [id, status, verification_url, session_id, expires_at]
      properties:
        id:
          type: string
        status:
          type: string
        verification_url:
          type: string
          format: uri
        session_id:
          type: string
        session_token:
          type: string
          description: Raw session token returned only when issuing a fresh hosted link.
        expires_at:
          type: string
          format: date-time
    VerificationSummary:
      type: object
      required:
        [id, status, purpose, external_reference, subject, policy, created_at]
      properties:
        id:
          type: string
        status:
          type: string
        purpose:
          type: string
        external_reference:
          type: string
        subject:
          type: object
          properties:
            id:
              type: string
            full_name:
              type: string
            email:
              type: string
        policy:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            version:
              type: integer
        created_at:
          type: string
          format: date-time
        completed_at:
          type: [string, "null"]
          format: date-time
...
    VerificationResult:
      type: object
      additionalProperties: false
      required:
        - schema_version
        - verification_id
        - status
        - decision
        - policy
        - workflow
        - check_provenance
        - timestamps
      properties:
        schema_version:
          type: string
          const: "1"
        verification_id:
          type: string
        status:
          type: string
        decision:
          type: [object, "null"]
          additionalProperties: false
          required:
            - outcome
            - type
            - reason_codes
            - approval_status
            - contract_version
            - decided_at
          properties:
            outcome:
              type: string
            type:
              type: string
            reason_codes:
              type: array
              items:
                type: string
            approval_status:
              type: string
            contract_version:
              type: string
            decided_at:
              type: string
              format: date-time
        policy:
          type: object
          additionalProperties: false
          required: [policy_id, version]
          properties:
            policy_id:
              type: [string, "null"]
            version:
              type: [integer, "null"]
        workflow:
          type: object
          additionalProperties: false
          required: [workflow_id, version_id, version]
          properties:
            workflow_id:
              type: [string, "null"]
            version_id:
              type: [string, "null"]
            version:
              type: [integer, "null"]
        check_provenance:
          type: array
          items:
            type: object
            additionalProperties: false
            required:
              - check_id
              - capability
              - status
              - provider
              - capability_contract_version
              - evidence
              - error_code
              - started_at
              - completed_at
              - duration_ms
            properties:
              check_id:
                type: string
              capability:
                type: string
              status:
                type: string
              provider:
                type: object
                additionalProperties: false
                required: [provider_id, code, type]
                properties:
                  provider_id:
                    type: string
                  code:
                    type: string
                  type:
                    type: string
              capability_contract_version:
                type: [string, "null"]
              evidence:
                type: object
                additionalProperties: false
                properties:
                  confidence_level:
                    type: string
                  confidence_score:
                    type: number
                  score:
                    type: number
                  match_score:
                    type: number
                  quality_score:
                    type: number
                  threshold_used:
                    type: number
                  model:
                    type: object
                    additionalProperties: false
                    required: [name, version]
                    properties:
                      name:
                        type: string
                      version:
                        type: string
                  issues:
                    type: array
                    items:
                      type: string
              error_code:
                type: [string, "null"]
              started_at:
                type: string
                format: date-time
              completed_at:
                type: [string, "null"]
                format: date-time
              duration_ms:
                type: [integer, "null"]
        timestamps:
          type: object
          additionalProperties: false
          required: [created_at, updated_at, completed_at, expires_at]
          properties:
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
            completed_at:
              type: [string, "null"]
              format: date-time
            expires_at:
              type: string
              format: date-time
    CursorPagination:
      type: object
      description: Cursor pagination metadata. Cursors are opaque and must be reused with the same filters.
      required: [limit, next_cursor, has_more]
      properties:
        limit:
          type: integer
          minimum: 1
          maximum: 100
        next_cursor:
          type: [string, "null"]
        has_more:
          type: boolean
...
    PortalUploadCreateRequest:
      type: object
      required: [purpose, mime_type, file_size_bytes]
      properties:
        purpose:
          type: string
          enum: [document_capture, selfie_capture, liveness_capture]
        mime_type:
          type: string
        file_size_bytes:
          type: integer
          minimum: 1
    PortalUploadCreateResponse:
      type: object
      required:
        [
          upload_id,
          upload_url,
          upload_headers,
          upload_transfer_path,
          expires_at,
        ]
      properties:
        upload_id:
          type: string
        upload_url:
          type: string
          format: uri
        upload_headers:
          type: object
          additionalProperties:
            type: string
        upload_transfer_path:
          type: string
        expires_at:
          type: string
          format: date-time
...
    OrganizationProfile:
      type: object
      required: [id, name, slug, status, settings, sandbox_usage]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        industry:
          type: string
        status:
          type: string
        tenant_id:
          type: string
        tenant_name:
          type: string
        tenant_status:
          type: string
        default_country_profile_id:
          type: string
        default_jurisdiction_id:
          type: string
        settings:
          type: object
          additionalProperties: true
        sandbox_usage:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    APIClientCreateResponse:
      allOf:
        - $ref: "#/components/schemas/APIClientSummary"
        - type: object
          required: [client_secret]
          properties:
            client_secret:
              type: string
...
  responses:
    ValidationError:
      description: Validation error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    PermissionDenied:
      description: Authentication failed or scope is missing.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: Resource was not found in the authenticated tenant.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
paths:
...
  /verifications/:
    get:
      summary: List verifications.
      parameters:
        - $ref: "#/components/parameters/ClientIdHeader"
        - name: status
          in: query
          schema:
            type: string
        - name: external_reference
          in: query
          schema:
            type: string
        - name: cursor
          in: query
          schema:
            type: string
          description: Opaque cursor returned by the previous page; reuse it with unchanged filters.
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Verification page.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    type: object
                    properties:
                      results:
                        type: array
                        items:
                          $ref: "#/components/schemas/VerificationSummary"
                      pagination:
                        $ref: "#/components/schemas/CursorPagination"
                  request_id:
                    type: string
        "403":
          $ref: "#/components/responses/PermissionDenied"
    post:
      summary: Create a hosted verification request.
      description: Requires `verifications:create`; API-client requests must include an active `policy_id` and an idempotency key.
      parameters:
        - $ref: "#/components/parameters/ClientIdHeader"
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerificationCreateRequest"
      responses:
        "201":
          description: Created verification.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    $ref: "#/components/schemas/VerificationCreateResponse"
                  request_id:
                    type: string
        "400":
          $ref: "#/components/responses/ValidationError"
        "403":
          $ref: "#/components/responses/PermissionDenied"
        "409":
          description: The idempotency key was previously used with a different payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /verifications/{verification_id}/result:
    get:
      summary: Retrieve the stable versioned verification result.
      parameters:
        - $ref: "#/components/parameters/ClientIdHeader"
        - name: verification_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Versioned verification result with allowlisted evidence lineage.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    $ref: "#/components/schemas/VerificationResult"
                  request_id:
                    type: string
        "403":
          $ref: "#/components/responses/PermissionDenied"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /uploads/:
    post:
      summary: Create a temporary upload transfer.
      description: Workspace-authenticated users create uploads before transferring file bytes from the browser or app.
      security:
        - platformUserSession: []
        - platformUserBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortalUploadCreateRequest"
      responses:
        "201":
          description: Upload transfer details.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    $ref: "#/components/schemas/PortalUploadCreateResponse"
                  request_id:
                    type: string
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/PermissionDenied"
        "403":
          $ref: "#/components/responses/PermissionDenied"
...
  /organization/me/:
    get:
      summary: Get the current organization profile.
      description: Returns the authenticated workspace's organization settings and sandbox usage.
      security:
        - platformUserSession: []
        - platformUserBearer: []
      responses:
        "200":
          description: Organization profile.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    $ref: "#/components/schemas/OrganizationProfile"
                  request_id:
                    type: string
        "401":
          $ref: "#/components/responses/PermissionDenied"
        "403":
          $ref: "#/components/responses/PermissionDenied"
    patch:
      summary: Update organization branding settings.
      security:
        - platformUserSession: []
        - platformUserBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                logo_storage_key:
                  type: string
                branding_image_storage_keys:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Updated organization profile.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    $ref: "#/components/schemas/OrganizationProfile"
                  request_id:
                    type: string
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/PermissionDenied"
        "403":
          $ref: "#/components/responses/PermissionDenied"
...
  /organization/me/verification-documents/{document_id}/content/:
    put:
      summary: Transfer supporting document bytes.
      description: Upload the PDF bytes for an initiated organization supporting document.
      security:
        - platformUserSession: []
        - platformUserBearer: []
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
        - name: Content-Type
          in: header
          required: true
          schema:
            const: application/pdf
      requestBody:
        required: true
        content:
          application/pdf:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Supporting document bytes stored.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    type: object
                    properties:
                      document_id:
                        type: string
                      status:
                        const: stored
                  request_id:
                    type: string
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/PermissionDenied"
        "403":
          $ref: "#/components/responses/PermissionDenied"
        "404":
          $ref: "#/components/responses/NotFound"
        "415":
          description: The uploaded document is not a PDF.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /organization/me/suspend:
    post:
      summary: Suspend the current workspace.
      description: Requires the exact organization name as confirmation and disables active workspace credentials and sessions.
      security:
        - platformUserSession: []
        - platformUserBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [confirmation]
              properties:
                confirmation:
                  type: string
      responses:
        "200":
          description: Workspace suspended.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    type: object
                    required: [suspended]
                    properties:
                      suspended:
                        const: true
                  request_id:
                    type: string
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/PermissionDenied"
        "403":
          $ref: "#/components/responses/PermissionDenied"
  /projects/:
    get:
      summary: List projects.
      security:
        - platformUserSession: []
        - platformUserBearer: []
      responses:
        "200":
          description: Project list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    type: object
                    properties:
                      results:
                        type: array
                        items:
                          $ref: "#/components/schemas/ProjectSummary"
                  request_id:
                    type: string
        "401":
          $ref: "#/components/responses/PermissionDenied"
        "403":
          $ref: "#/components/responses/PermissionDenied"
    post:
      summary: Create a project.
      security:
        - platformUserSession: []
        - platformUserBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectCreateRequest"
      responses:
        "201":
          description: Created project.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    $ref: "#/components/schemas/ProjectSummary"
                  request_id:
                    type: string
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/PermissionDenied"
        "403":
          $ref: "#/components/responses/PermissionDenied"
...
  /projects/{project_id}/workflows:instantiate:
    post:
      summary: Instantiate a workflow template in a project.
      description: Creates an idempotent workflow from a published template. Production projects require an approved organization.
      security:
        - platformUserSession: []
        - platformUserBearer: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            maxLength: 255
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [template_id, template_version]
              properties:
                template_id:
                  type: string
                template_version:
                  type: string
                name:
                  type: string
      responses:
        "201":
          description: Workflow instantiated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    $ref: "#/components/schemas/WorkflowSummary"
                  request_id:
                    type: string
        "200":
          description: Existing idempotent workflow returned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    $ref: "#/components/schemas/WorkflowSummary"
                  request_id:
                    type: string
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/PermissionDenied"
        "403":
          $ref: "#/components/responses/PermissionDenied"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Idempotency key conflicts with a previous request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api-clients/:
    get:
      summary: List API clients.
      security:
        - platformUserSession: []
        - platformUserBearer: []
      responses:
        "200":
          description: API client list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    type: object
                    properties:
                      results:
                        type: array
                        items:
                          $ref: "#/components/schemas/APIClientSummary"
                  request_id:
                    type: string
        "401":
          $ref: "#/components/responses/PermissionDenied"
        "403":
          $ref: "#/components/responses/PermissionDenied"
    post:
      summary: Create an API client.
      description: Pending workspaces can use this route to mint a sandbox test key for API testing.
      security:
        - platformUserSession: []
        - platformUserBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/APIClientCreateRequest"
      responses:
        "201":
          description: Created API client.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    $ref: "#/components/schemas/APIClientCreateResponse"
                  request_id:
                    type: string
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/PermissionDenied"
        "403":
          $ref: "#/components/responses/PermissionDenied"
...
  /webhook-endpoints/:
    get:
      summary: List webhook endpoints.
      security:
        - platformUserSession: []
        - platformUserBearer: []
      responses:
        "200":
          description: Webhook endpoint list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    type: object
                    properties:
                      results:
                        type: array
                        items:
                          $ref: "#/components/schemas/WebhookEndpointSummary"
                  request_id:
                    type: string
        "401":
          $ref: "#/components/responses/PermissionDenied"
        "403":
          $ref: "#/components/responses/PermissionDenied"
    post:
      summary: Create a webhook endpoint.
      description: Pending workspaces can create a single disabled sandbox webhook for test delivery.
      security:
        - platformUserSession: []
        - platformUserBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookEndpointCreateRequest"
      responses:
        "201":
          description: Created webhook endpoint.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    $ref: "#/components/schemas/WebhookEndpointCreateResponse"
                  request_id:
                    type: string
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/PermissionDenied"
        "403":
          $ref: "#/components/responses/PermissionDenied"
...
  /verifications/manual-reviews:
    get:
      summary: List verifications requiring manual review.
      security:
        - platformUserSession: []
        - platformUserBearer: []
      responses:
        "200":
          description: Manual review queue.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    type: object
                    properties:
                      results:
                        type: array
                        items:
                          $ref: "#/components/schemas/ManualReviewSummary"
                      pagination:
                        $ref: "#/components/schemas/PagePagination"
                  request_id:
                    type: string
        "401":
          $ref: "#/components/responses/PermissionDenied"
        "403":
          $ref: "#/components/responses/PermissionDenied"
...
  /verifications/{verification_id}/evidence-report/download.pdf:
    get:
      summary: Download evidence-report PDF.
      security:
        - platformUserSession: []
        - platformUserBearer: []
      parameters:
        - name: verification_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Evidence report PDF download.
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/PermissionDenied"
        "403":
          $ref: "#/components/responses/PermissionDenied"
        "404":
          $ref: "#/components/responses/NotFound"